pyproject.toml
[tool.poetry]
name = "bain-test-alvaro-marques"
version = "0.1.0"
description = "Property valuation model API for Chilean real estate"
authors = ["AlvaroMarquesAndrade <alvaro.marques.andrade@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pandas = "^2.1.3"
numpy = "^1.24.3"
scikit-learn = "^1.3.2"
xgboost = "^2.0.2"
joblib = "^1.3.2"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
python-multipart = "^0.0.6"
requests = "^2.32.5"
category-encoders = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"
black = "^23.11.0"
flake8 = "^6.1.0"
ipython = "^8.17.2"

[tool.poetry.scripts]
train = "scripts.run_training:main"
api = "src.api.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"